protocol: add targeted queue delivery
Bump the protocol version and widen message user data to 64 bits so a
producer can attach generation and epoch identity to each delivery.

Assign a unique ID to every client in a host session and carry that ID
with client control messages. Validate the source against the current
subscription before exposing control data to the host, preventing stale
messages from disconnected clients from being accepted.

Add targeted queue posting, active client enumeration, and separate
queries for payload ownership and outstanding recipient state. Leave
payload ownership with the caller when no targeted recipient matches.

Update client queue traversal to handle sparse targeted messages and to
safely advance, release, unsubscribe, and reuse subscriber slots.

Invalidate timed-out clients across every queue. Reclaim acknowledged
tail entries behind a live queue head so targeted traffic is not blocked
by an unrelated slow subscriber, while retaining payload ownership
until every live recipient releases it.

Keep the expanded shared structures explicitly aligned and update the
test client for 64-bit message user data.

+++ lgmp/include/lgmp/client.h
@@ -44,7 +44,7 @@ LGMP_STATUS lgmpClientUnsubscribe(PLGMPClientQueue * result);
 
 typedef struct
 {
-  uint32_t   udata;
+  uint64_t   udata;
   uint32_t   size;
   void     * mem;
 }
